ayncwrite for xskstream.
There is a bug somewere (posibly in libbpf/kernel) or i dont yet
fully understand how the fillq works. It looks like if we first
fill the queue (with all packets set with with_fill_size, we
can only receive that number of packets).

After a first investigation, it looks like q->cached_cons is
incremented locally, and then written at some point. It may
be that if the queue is full, the actual write will never
take place.